json: Return wrapper object in streaming parsing
