+++ modules/json/src/index.js
@@ -1,7 +1,1 @@
 export {JSONLoader} from './json-loader';
-
-export {default as JSONParser} from './lib/json-parser';
-export {default as StreamingJSONParser} from './lib/streaming-json-parser';
-
-// Fork of clarinet, these are considered experimental exports
-export {default as ClarinetParser} from './clarinet/clarinet';
