+++ modules/json/test/lib/clarinet/clarinet.bench.js
@@ -1,4 +1,4 @@
-import {ClarinetParser} from '@loaders.gl/json';
+import ClarinetParser from '@loaders.gl/json/lib/clarinet/clarinet';
 
 import BASIC from '../data/clarinet/basic.json';
